+++ CHANGELOG.md
@@ -71,6 +71,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Relationship facet in Filter nomenclature
 
 ### Changed
+- Updated "ICZN binomial" to "ICZN binominal" following the Code dictionary
 - DwcOccurrence is rebuilt/refreshed each time Browse Collection Object is hit
 - `footprintWKT` is hidden in Browse CollectionOjbect [#2559]
 - Tweak geo coordinate matching on verbatim labels
@@ -729,7 +730,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Fixed
 
-- updated softvalidation for non binomial names
+- updated softvalidation for non binominal names
 - updated label for species-group rank TaxonName
 - Compute print column divisions with barcode style labels [#1993]
 - Object tag for TaxonNameRelationship inverted [#2100]